Split draw_border_square into two separate functions
Now there is draw_square and draw_board_square. draw_board_square takes
a position in the game array and prints on that portion of the screen.
draw_square is a generalized one that just draws a square anywhere on the screen.

Used this to implement a visualization of the held block on the left
padding.